dcerpc: fix compiler warning
git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@53514 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ dcerpc/idl_compiler/getflags.c
@@ -168,7 +168,7 @@ boolean is_number
         str++;
 
     for ( ; *str != '\0' ; str++)
-        if (!isdigit(*str))
+        if (!isdigit((int)*str))
             return false;
 
     return true;
